+++ modules/apps/users-admin/users-admin-test/src/testIntegration/java/com/liferay/users/admin/service/user/test/UserServiceWhenAddingOrRemovingPasswordPolicyUsersTest.java
@@ -12,8 +12,9 @@
  * details.
  */
 
-package com.liferay.portal.service.user;
+package com.liferay.users.admin.service.user.test;
 
+import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.portal.kernel.model.PasswordPolicy;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.service.PasswordPolicyLocalServiceUtil;
@@ -32,12 +33,14 @@ import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 
 /**
  * @author Brian Wing Shun Chan
  * @author José Manuel Navarro
  * @author Drew Brokke
  */
+@RunWith(Arquillian.class)
 public class UserServiceWhenAddingOrRemovingPasswordPolicyUsersTest {
 
 	@ClassRule
