+++ modules/apps/users-admin/users-admin-test/src/testIntegration/java/com/liferay/users/admin/service/user/test/UserServiceWhenAddingUserWithSpecialCharactersScreenNameTest.java
@@ -12,8 +12,9 @@
  * details.
  */
 
-package com.liferay.portal.service.user;
+package com.liferay.users.admin.service.user.test;
 
+import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.security.auth.DefaultScreenNameValidator;
@@ -34,12 +35,14 @@ import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 
 /**
  * @author Brian Wing Shun Chan
  * @author José Manuel Navarro
  * @author Drew Brokke
  */
+@RunWith(Arquillian.class)
 public class UserServiceWhenAddingUserWithSpecialCharactersScreenNameTest {
 
 	@ClassRule
