+++ modules/apps/users-admin/users-admin-test/src/testIntegration/java/com/liferay/users/admin/service/user/test/UserServiceWhenGroupAdminUnsetsGroupUsersTest.java
@@ -12,14 +12,16 @@
  * details.
  */
 
-package com.liferay.portal.service.user;
+package com.liferay.users.admin.service.user.test;
 
+import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.Organization;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.service.UserLocalServiceUtil;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.DeleteAfterTestRun;
+import com.liferay.portal.kernel.test.util.GroupTestUtil;
 import com.liferay.portal.kernel.test.util.OrganizationTestUtil;
 import com.liferay.portal.kernel.test.util.UserTestUtil;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
@@ -29,13 +31,15 @@ import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 
 /**
  * @author Brian Wing Shun Chan
  * @author José Manuel Navarro
  * @author Drew Brokke
  */
-public class UserServiceOrganizationOwnerUnsetsUsersForSiteOrganizationTest {
+@RunWith(Arquillian.class)
+public class UserServiceWhenGroupAdminUnsetsGroupUsersTest {
 
 	@ClassRule
 	@Rule
@@ -46,21 +50,20 @@ public class UserServiceOrganizationOwnerUnsetsUsersForSiteOrganizationTest {
 	public void setUp() throws Exception {
 		_organization = OrganizationTestUtil.addOrganization(true);
 
-		_group = _organization.getGroup();
+		_group = GroupTestUtil.addGroup();
 
-		_organizationOwnerUser = UserTestUtil.addOrganizationOwnerUser(
-			_organization);
+		_groupAdminUser = UserTestUtil.addGroupAdminUser(_group);
 	}
 
 	@Test
-	public void testShouldUnsetSiteAdmin() throws Exception {
+	public void testShouldUnsetGroupAdmin() throws Exception {
 		User groupAdminUser = UserTestUtil.addGroupAdminUser(_group);
 
 		try {
 			UserServiceTestUtil.unsetGroupUsers(
-				_group.getGroupId(), _organizationOwnerUser, groupAdminUser);
+				_group.getGroupId(), _groupAdminUser, groupAdminUser);
 
-			Assert.assertFalse(
+			Assert.assertTrue(
 				UserLocalServiceUtil.hasGroupUser(
 					_group.getGroupId(), groupAdminUser.getUserId()));
 		}
@@ -70,14 +73,14 @@ public class UserServiceOrganizationOwnerUnsetsUsersForSiteOrganizationTest {
 	}
 
 	@Test
-	public void testShouldUnsetSiteOwner() throws Exception {
+	public void testShouldUnsetGroupOwner() throws Exception {
 		User groupOwnerUser = UserTestUtil.addGroupOwnerUser(_group);
 
 		try {
 			UserServiceTestUtil.unsetGroupUsers(
-				_group.getGroupId(), _organizationOwnerUser, groupOwnerUser);
+				_group.getGroupId(), _groupAdminUser, groupOwnerUser);
 
-			Assert.assertFalse(
+			Assert.assertTrue(
 				UserLocalServiceUtil.hasGroupUser(
 					_group.getGroupId(), groupOwnerUser.getUserId()));
 		}
@@ -86,12 +89,53 @@ public class UserServiceOrganizationOwnerUnsetsUsersForSiteOrganizationTest {
 		}
 	}
 
+	@Test
+	public void testShouldUnsetOrganizationAdmin() throws Exception {
+		User organizationAdminUser = UserTestUtil.addOrganizationAdminUser(
+			_organization);
+
+		try {
+			UserServiceTestUtil.unsetOrganizationUsers(
+				_organization.getOrganizationId(), _groupAdminUser,
+				organizationAdminUser);
+
+			Assert.assertTrue(
+				UserLocalServiceUtil.hasOrganizationUser(
+					_organization.getOrganizationId(),
+					organizationAdminUser.getUserId()));
+		}
+		finally {
+			UserLocalServiceUtil.deleteUser(organizationAdminUser);
+		}
+	}
+
+	@Test
+	public void testShouldUnsetOrganizationOwner() throws Exception {
+		User organizationOwnerUser = UserTestUtil.addOrganizationOwnerUser(
+			_organization);
+
+		try {
+			UserServiceTestUtil.unsetOrganizationUsers(
+				_organization.getOrganizationId(), _groupAdminUser,
+				organizationOwnerUser);
+
+			Assert.assertTrue(
+				UserLocalServiceUtil.hasOrganizationUser(
+					_organization.getOrganizationId(),
+					organizationOwnerUser.getUserId()));
+		}
+		finally {
+			UserLocalServiceUtil.deleteUser(organizationOwnerUser);
+		}
+	}
+
+	@DeleteAfterTestRun
 	private Group _group;
 
 	@DeleteAfterTestRun
-	private Organization _organization;
+	private User _groupAdminUser;
 
 	@DeleteAfterTestRun
-	private User _organizationOwnerUser;
+	private Organization _organization;
 
 }
